+++ src/enterprise-operations/service-registration.js
@@ -1,4 +1,5 @@
 import { ServiceCollection } from '../foundation/di/service-collection.js';
+import { OperationsCommandControlDescriptor } from './command-control/operations-command-control-descriptor.js';
 import { EnterpriseOperationsDescriptor } from './enterprise-operations-descriptor.js';
 import { EnterpriseOperatingModelDescriptor } from './operating-model/enterprise-operating-model-descriptor.js';
 import { EnterpriseOperationsOverviewDescriptor } from './overview/enterprise-operations-overview-descriptor.js';
@@ -11,6 +12,7 @@ export function addEnterpriseOperations(services) {
   services.registerSingleton('EnterpriseOperationsDescriptor', () => new EnterpriseOperationsDescriptor());
   services.registerSingleton('EnterpriseOperationsOverviewDescriptor', () => new EnterpriseOperationsOverviewDescriptor());
   services.registerSingleton('EnterpriseOperatingModelDescriptor', () => new EnterpriseOperatingModelDescriptor());
+  services.registerSingleton('OperationsCommandControlDescriptor', () => new OperationsCommandControlDescriptor());
 
   return services;
 }